pyproject.toml
[project]
name = "citations"
version = "0.1.0"
description = "Citation discovery and wiki sync for open-source research tools"
requires-python = ">=3.10"
dependencies = [
    "bibtexparser>=1.4.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
